feature: del log and re-mig on welog module all
