--- log.go
@@ -1,34 +1,0 @@
-package welog
-
-import (
-	"fmt"
-	"time"
-)
-
-func (l *Logger) log(level LogLevel, message string) {
-	if l.level >= level {
-		timestamp := time.Now().Format(TimestampFormat)
-		levelString := LevelToString(level)
-		colorFunc := ColorType(level)
-
-		logMessage := fmt.Sprintf("[%s] [%s] %s\n", timestamp, colorFunc(levelString), message)
-
-		fmt.Fprint(l.writer, logMessage)
-	}
-}
-
-func (l *Logger) Error(message string) {
-	l.log(Error, message)
-}
-
-func (l *Logger) Warn(message string) {
-	l.log(Warn, message)
-}
-
-func (l *Logger) Info(message string) {
-	l.log(Info, message)
-}
-
-func (l *Logger) Debug(message string) {
-	l.log(Debug, message)
-}
